+++ features/config/main_branch/display.feature
@@ -1,8 +1,9 @@
+@this
 Feature: display the main branch configuration
 
   Scenario: not configured
     Given the main branch is not set
-    When I run "git-town main-branch"
+    When I run "git-town config main-branch"
     Then it prints:
       """
       [none]
@@ -10,7 +11,7 @@ Feature: display the main branch configuration
 
   Scenario: configured
     Given the main branch is "main"
-    When I run "git-town main-branch"
+    When I run "git-town config main-branch"
     Then it prints:
       """
       main
